Created action to auto-deploy mkdocs site
This means that when pull requests go through to `main` that change the contents or configuration for the site, these changes will automatically deploy and become visible in the public view.

Note: This is untested. I may make a bunch of commits after this trying to tinker until I get it right.